+++ Source/OpenLight/Texture/SolidNoise.h
@@ -11,19 +11,19 @@ public:
 	SolidNoise();
 
 public:
-	double noise( const Vector3f& ) const;
+	float noise( const Vector3f& ) const;
 
-	double turbulence( const Vector3f& p , int depth ) const;
+	float turbulence( const Vector3f& p , int depth ) const;
 
-	double dturbulence( const Vector3f& p , int depth , double d ) const;
+	float dturbulence( const Vector3f& p , int depth , float d ) const;
 
-	double omega( double t ) const;
+	float omega( float t ) const;
 
 	Vector3f gamma( int i , int j , int k ) const;
 
 	int intGamma( int i , int j ) const;
 
-	double knot( int i , int j , int k , Vector3f& v ) const;
+	float knot( int i , int j , int k , Vector3f& v ) const;
 
 public:
 	Vector3f grad[16];
